Factor Flag out of Distribution.Simple.Setup
